Fixed a few routines.

+++ src/multi_bspline_eval_std_s.h
@@ -40,7 +40,7 @@ eval_multi_UBspline_1d_s (multi_UBspline_1d_s *spline,
   x -= spline->x_grid.start;
   float ux = x*spline->x_grid.delta_inv;
   float ipartx, tx;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
   
   float tpx[4], a[4];
   tpx[0] = tx*tx*tx;  tpx[1] = tx*tx;  tpx[2] = tx;  tpx[3] = 1.0;
@@ -74,7 +74,7 @@ eval_multi_UBspline_1d_s_vg (multi_UBspline_1d_s *spline,
   x -= spline->x_grid.start;
   float ux = x*spline->x_grid.delta_inv;
   float ipartx, tx;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
   
   float tpx[4], a[4], da[4];
   tpx[0] = tx*tx*tx;  tpx[1] = tx*tx;  tpx[2] = tx;  tpx[3] = 1.0;
@@ -120,7 +120,7 @@ eval_multi_UBspline_1d_s_vgl (multi_UBspline_1d_s *spline,
   x -= spline->x_grid.start;
   float ux = x*spline->x_grid.delta_inv;
   float ipartx, tx;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
   
   float tpx[4], a[4], da[4], d2a[4];
   tpx[0] = tx*tx*tx;  tpx[1] = tx*tx;  tpx[2] = tx;  tpx[3] = 1.0;
@@ -176,8 +176,8 @@ eval_multi_UBspline_2d_s (multi_UBspline_2d_s *spline,
   float ux = x*spline->x_grid.delta_inv;
   float uy = y*spline->y_grid.delta_inv;
   float ipartx, iparty, tx, ty;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
-  ty = modf (uy, &iparty);  int iy = (int) iparty;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
+  ty = modff (uy, &iparty);  int iy = (int) iparty;
   
   float tpx[4], tpy[4], a[4], b[4];
   tpx[0] = tx*tx*tx;  tpx[1] = tx*tx;  tpx[2] = tx;  tpx[3] = 1.0;
@@ -221,8 +221,8 @@ eval_multi_UBspline_2d_s_vg (multi_UBspline_2d_s *spline,
   float ux = x*spline->x_grid.delta_inv;
   float uy = y*spline->y_grid.delta_inv;
   float ipartx, iparty, tx, ty;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
-  ty = modf (uy, &iparty);  int iy = (int) iparty;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
+  ty = modff (uy, &iparty);  int iy = (int) iparty;
   
   float tpx[4], tpy[4], a[4], b[4], da[4], db[4];
   tpx[0] = tx*tx*tx;  tpx[1] = tx*tx;  tpx[2] = tx;  tpx[3] = 1.0;
@@ -290,8 +290,8 @@ eval_multi_UBspline_2d_s_vgl (multi_UBspline_2d_s *spline,
   float ux = x*spline->x_grid.delta_inv;
   float uy = y*spline->y_grid.delta_inv;
   float ipartx, iparty, tx, ty;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
-  ty = modf (uy, &iparty);  int iy = (int) iparty;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
+  ty = modff (uy, &iparty);  int iy = (int) iparty;
   
   float tpx[4], tpy[4], a[4], b[4], da[4], db[4], d2a[4], d2b[4];
   tpx[0] = tx*tx*tx;  tpx[1] = tx*tx;  tpx[2] = tx;  tpx[3] = 1.0;
@@ -379,8 +379,8 @@ eval_multi_UBspline_2d_s_vgh (multi_UBspline_2d_s *spline,
   float ux = x*spline->x_grid.delta_inv;
   float uy = y*spline->y_grid.delta_inv;
   float ipartx, iparty, tx, ty;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
-  ty = modf (uy, &iparty);  int iy = (int) iparty;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
+  ty = modff (uy, &iparty);  int iy = (int) iparty;
   
   float tpx[4], tpy[4], a[4], b[4], da[4], db[4], d2a[4], d2b[4];
   tpx[0] = tx*tx*tx;  tpx[1] = tx*tx;  tpx[2] = tx;  tpx[3] = 1.0;
@@ -475,9 +475,9 @@ eval_multi_UBspline_3d_s (multi_UBspline_3d_s *spline,
   float uy = y*spline->y_grid.delta_inv;
   float uz = z*spline->z_grid.delta_inv;
   float ipartx, iparty, ipartz, tx, ty, tz;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
-  ty = modf (uy, &iparty);  int iy = (int) iparty;
-  tz = modf (uz, &ipartz);  int iz = (int) ipartz;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
+  ty = modff (uy, &iparty);  int iy = (int) iparty;
+  tz = modff (uz, &ipartz);  int iz = (int) ipartz;
   
   float tpx[4], tpy[4], tpz[4], a[4], b[4], c[4];
   tpx[0] = tx*tx*tx;  tpx[1] = tx*tx;  tpx[2] = tx;  tpx[3] = 1.0;
@@ -531,9 +531,9 @@ eval_multi_UBspline_3d_s_vg (multi_UBspline_3d_s *spline,
   float uy = y*spline->y_grid.delta_inv;
   float uz = z*spline->z_grid.delta_inv;
   float ipartx, iparty, ipartz, tx, ty, tz;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
-  ty = modf (uy, &iparty);  int iy = (int) iparty;
-  tz = modf (uz, &ipartz);  int iz = (int) ipartz;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
+  ty = modff (uy, &iparty);  int iy = (int) iparty;
+  tz = modff (uz, &ipartz);  int iz = (int) ipartz;
   
   float tpx[4], tpy[4], tpz[4], a[4], b[4], c[4], 
     da[4], db[4], dc[4];
@@ -622,9 +622,9 @@ eval_multi_UBspline_3d_s_vgl (multi_UBspline_3d_s *spline,
   float uy = y*spline->y_grid.delta_inv;
   float uz = z*spline->z_grid.delta_inv;
   float ipartx, iparty, ipartz, tx, ty, tz;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
-  ty = modf (uy, &iparty);  int iy = (int) iparty;
-  tz = modf (uz, &ipartz);  int iz = (int) ipartz;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
+  ty = modff (uy, &iparty);  int iy = (int) iparty;
+  tz = modff (uz, &ipartz);  int iz = (int) ipartz;
   
   float tpx[4], tpy[4], tpz[4], a[4], b[4], c[4], 
     da[4], db[4], dc[4], d2a[4], d2b[4], d2c[4];
@@ -737,9 +737,9 @@ eval_multi_UBspline_3d_s_vgh (multi_UBspline_3d_s *spline,
   float uy = y*spline->y_grid.delta_inv;
   float uz = z*spline->z_grid.delta_inv;
   float ipartx, iparty, ipartz, tx, ty, tz;
-  tx = modf (ux, &ipartx);  int ix = (int) ipartx;
-  ty = modf (uy, &iparty);  int iy = (int) iparty;
-  tz = modf (uz, &ipartz);  int iz = (int) ipartz;
+  tx = modff (ux, &ipartx);  int ix = (int) ipartx;
+  ty = modff (uy, &iparty);  int iy = (int) iparty;
+  tz = modff (uz, &ipartz);  int iz = (int) ipartz;
   
   float tpx[4], tpy[4], tpz[4], a[4], b[4], c[4], 
     da[4], db[4], dc[4], d2a[4], d2b[4], d2c[4];
